Base/requirements.txt
pyserial>=3.5
PyYAML>=6.0.1

